MOD: Improved the performance of the codebase by storing the positions into halfs instead of floats. RESULTS: The performance improved from 4.3 to 3.4ms per tick, very good improvements for something this late in the game.

+++ Source/game.h
@@ -9,8 +9,8 @@
 typedef struct
 {
 	int16_t spriteIndicesAndScales[GAME_MAX_INSTANCE_COUNT * 2];
-	float positionX[GAME_MAX_INSTANCE_COUNT];
-	float positionY[GAME_MAX_INSTANCE_COUNT];
+	uint16_t positionX[GAME_MAX_INSTANCE_COUNT];
+	uint16_t positionY[GAME_MAX_INSTANCE_COUNT];
 } Game_InstanceBuffer;
 
 void game_init(Game_InstanceBuffer* buffer);
